even more generator refactoring to files
