+++ src/stan/lang/generator.hpp
@@ -12,8 +12,18 @@
 
 #include <stan/lang/generator/expression_visgen.hpp>
 #include <stan/lang/generator/printable_visgen.hpp>
+#include <stan/lang/generator/idx_visgen.hpp>
+#include <stan/lang/generator/idx_user_visgen.hpp>
+#include <stan/lang/generator/init_local_var_visgen.hpp>
 #include <stan/lang/generator/init_vars_visgen.hpp>
+#include <stan/lang/generator/local_var_decl_visgen.hpp>
 #include <stan/lang/generator/local_var_init_nan_visgen.hpp>
+#include <stan/lang/generator/member_var_decl_visgen.hpp>
+#include <stan/lang/generator/statement_visgen.hpp>
+#include <stan/lang/generator/validate_var_decl_visgen.hpp>
+#include <stan/lang/generator/validate_transformed_params_visgen.hpp>
+#include <stan/lang/generator/var_resizing_visgen.hpp>
+#include <stan/lang/generator/var_size_validating_visgen.hpp>
 #include <stan/lang/generator/visgen.hpp>
 
 #include <stan/lang/generator/constants.hpp>
@@ -23,15 +33,24 @@
 #include <stan/lang/generator/generate_comment.hpp>
 #include <stan/lang/generator/generate_eigen_index_expression.hpp>
 #include <stan/lang/generator/generate_expression.hpp>
+#include <stan/lang/generator/generate_idx.hpp>
+#include <stan/lang/generator/generate_idxs.hpp>
+#include <stan/lang/generator/generate_idxs_user.hpp>
 #include <stan/lang/generator/generate_indent.hpp>
 #include <stan/lang/generator/generate_include.hpp>
 #include <stan/lang/generator/generate_includes.hpp>
 #include <stan/lang/generator/generate_indexed_expr.hpp>
 #include <stan/lang/generator/generate_indexed_expr_user.hpp>
 #include <stan/lang/generator/generate_initializer.hpp>
+#include <stan/lang/generator/generate_initialization.hpp>
+#include <stan/lang/generator/generate_local_var_decls.hpp>
+#include <stan/lang/generator/generate_local_var_inits.hpp>
+#include <stan/lang/generator/generate_member_var_decls.hpp>
 #include <stan/lang/generator/generate_namespace_end.hpp>
 #include <stan/lang/generator/generate_namespace_start.hpp>
 #include <stan/lang/generator/generate_printable.hpp>
+#include <stan/lang/generator/generate_private_decl.hpp>
+#include <stan/lang/generator/generate_public_decl.hpp>
 #include <stan/lang/generator/generate_quoted_expression.hpp>
 #include <stan/lang/generator/generate_quoted_string.hpp>
 #include <stan/lang/generator/generate_real_var_type.hpp>
@@ -43,8 +62,13 @@
 #include <stan/lang/generator/generate_usings.hpp>
 #include <stan/lang/generator/generate_validate_context_size.hpp>
 #include <stan/lang/generator/generate_validate_positive.hpp>
+#include <stan/lang/generator/generate_validate_transformed_params.hpp>
+#include <stan/lang/generator/generate_validate_var_decl.hpp>
+#include <stan/lang/generator/generate_validate_var_decls.hpp>
+#include <stan/lang/generator/generate_var_resizing.hpp>
 #include <stan/lang/generator/generate_version_comment.hpp>
 #include <stan/lang/generator/generate_void_statement.hpp>
+
 
 
 #include <boost/variant/apply_visitor.hpp>
@@ -82,1480 +106,8 @@ namespace stan {
     void generate_idxs(const std::vector<idx>& idxs, std::ostream& o);
     void generate_idxs_user(const std::vector<idx>& idxs, std::ostream& o);
 
-    struct var_size_validating_visgen : public visgen {
-      const std::string stage_;
-      var_size_validating_visgen(std::ostream& o, const std::string& stage)
-        : visgen(o),
-          stage_(stage) {
-      }
-      void operator()(const nil& /*x*/) const { }  // dummy
-      void operator()(const int_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "int", x.dims_);
-      }
-      void operator()(const double_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "double", x.dims_);
-      }
-      void operator()(const vector_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "vector_d",
-                                       x.dims_, x.M_);
-      }
-      void operator()(const row_vector_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "row_vector_d",
-                                       x.dims_, x.N_);
-      }
-      void operator()(const unit_vector_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "vector_d",
-                                       x.dims_, x.K_);
-      }
-      void operator()(const simplex_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "vector_d",
-                                       x.dims_, x.K_);
-      }
-      void operator()(const ordered_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "vector_d",
-                                       x.dims_, x.K_);
-      }
-      void operator()(const positive_ordered_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "vector_d",
-                                       x.dims_, x.K_);
-      }
-      void operator()(const matrix_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "matrix_d",
-                                       x.dims_, x.M_, x.N_);
-      }
-      void operator()(const cholesky_factor_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "matrix_d",
-                                       x.dims_, x.M_, x.N_);
-      }
-      void operator()(const cholesky_corr_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "matrix_d",
-                                       x.dims_, x.K_, x.K_);
-      }
-      void operator()(const cov_matrix_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "matrix_d",
-                                       x.dims_, x.K_, x.K_);
-      }
-      void operator()(const corr_matrix_var_decl& x) const {
-        generate_validate_context_size(o_, stage_, x.name_, "matrix_d",
-                                       x.dims_, x.K_, x.K_);
-      }
-    };
 
 
-
-    /*
-     * generates variable declaration
-     */
-    void generate_initialization(std::ostream& o,
-                                 const std::string& var_name,
-                                 const std::string& base_type,
-                                 const std::vector<expression>& dims,
-                                 const expression& type_arg1 = expression(),
-                                 const expression& type_arg2 = expression()) {
-      // check array dimensions
-      for (size_t i = 0; i < dims.size(); ++i)
-        generate_validate_positive(var_name, dims[i], 2, o);
-      // check vector, row_vector, and matrix rows
-      if (!is_nil(type_arg1))
-        generate_validate_positive(var_name, type_arg1, 2, o);
-      // check matrix cols
-      if (!is_nil(type_arg2))
-        generate_validate_positive(var_name, type_arg2, 2, o);
-
-      // initialize variable
-      o << INDENT2
-        << var_name << " = ";
-      generate_type(base_type, dims, dims.size(), o);
-      generate_initializer(o, base_type, dims, type_arg1, type_arg2);
-    }
-
-    struct var_resizing_visgen : public visgen {
-      explicit var_resizing_visgen(std::ostream& o)
-        : visgen(o) {
-      }
-      void operator()(const nil& /*x*/) const { }  // dummy
-      void operator()(const int_var_decl& x) const {
-        generate_initialization(o_, x.name_, "int", x.dims_,
-                                nil(), nil());
-      }
-      void operator()(const double_var_decl& x) const {
-        generate_initialization(o_, x.name_, "double", x.dims_, nil(), nil());
-      }
-      void operator()(const vector_var_decl& x) const {
-        generate_initialization(o_, x.name_, "vector_d", x.dims_, x.M_, nil());
-      }
-      void operator()(const row_vector_var_decl& x) const {
-        generate_initialization(o_, x.name_, "row_vector_d", x.dims_, x.N_,
-                                nil());
-      }
-      void operator()(const unit_vector_var_decl& x) const {
-        generate_initialization(o_, x.name_, "vector_d", x.dims_, x.K_, nil());
-      }
-      void operator()(const simplex_var_decl& x) const {
-        generate_initialization(o_, x.name_, "vector_d", x.dims_, x.K_, nil());
-      }
-      void operator()(const ordered_var_decl& x) const {
-        generate_initialization(o_, x.name_, "vector_d", x.dims_, x.K_, nil());
-      }
-      void operator()(const positive_ordered_var_decl& x) const {
-        generate_initialization(o_, x.name_, "vector_d", x.dims_, x.K_, nil());
-      }
-      void operator()(const matrix_var_decl& x) const {
-        generate_initialization(o_, x.name_, "matrix_d",
-                                x.dims_, x.M_, x.N_);
-      }
-      void operator()(const cholesky_factor_var_decl& x) const {
-        generate_initialization(o_, x.name_, "matrix_d", x.dims_, x.M_, x.N_);
-      }
-      void operator()(const cholesky_corr_var_decl& x) const {
-        generate_initialization(o_, x.name_, "matrix_d", x.dims_, x.K_, x.K_);
-      }
-      void operator()(const cov_matrix_var_decl& x) const {
-        generate_initialization(o_, x.name_, "matrix_d", x.dims_, x.K_, x.K_);
-      }
-      void operator()(const corr_matrix_var_decl& x) const {
-        generate_initialization(o_, x.name_, "matrix_d", x.dims_, x.K_, x.K_);
-      }
-    };
-
-    void generate_var_resizing(const std::vector<var_decl>& vs,
-                               std::ostream& o) {
-      var_resizing_visgen vis_resizer(o);
-      init_vars_visgen vis_filler(2, o);
-      for (size_t i = 0; i < vs.size(); ++i) {
-        boost::apply_visitor(vis_resizer, vs[i].decl_);
-        boost::apply_visitor(vis_filler, vs[i].decl_);
-        if (vs[i].has_def()) {
-          o << INDENT2 << "stan::math::assign("
-            << vs[i].name()
-            << ",";
-          generate_expression(vs[i].def(), o);
-          o << ");" << EOL;
-        }
-      }
-    }
-
-    const std::vector<expression> EMPTY_EXP_VECTOR(0);
-
-    struct init_local_var_visgen : public visgen {
-      const bool declare_vars_;
-      const bool is_var_context_;
-      explicit init_local_var_visgen(bool declare_vars,
-                            bool is_var_context,
-                            std::ostream& o)
-        : visgen(o),
-          declare_vars_(declare_vars),
-          is_var_context_(is_var_context) {
-      }
-      template <typename D>
-      void generate_initialize_array_bounded(const D& x,
-                                             const std::string& base_type,
-                                             const std::string& read_fun_prefix,
-                                const std::vector<expression>& dim_args) const {
-        std::vector<expression> read_args;
-        std::string read_fun(read_fun_prefix);
-        if (has_lub(x)) {
-          read_fun += "_lub";
-          read_args.push_back(x.range_.low_);
-          read_args.push_back(x.range_.high_);
-        } else if (has_lb(x)) {
-          read_fun += "_lb";
-          read_args.push_back(x.range_.low_);
-        } else if (has_ub(x)) {
-          read_fun += "_ub";
-          read_args.push_back(x.range_.high_);
-        }
-        for (size_t i = 0; i < dim_args.size(); ++i)
-          read_args.push_back(dim_args[i]);
-        generate_initialize_array(base_type, read_fun, read_args,
-                                    x.name_, x.dims_);
-      }
-      void operator()(const nil& /*x*/) const { }
-      void operator()(const int_var_decl& x) const {
-        generate_initialize_array("int", "integer", EMPTY_EXP_VECTOR,
-                                  x.name_, x.dims_);
-      }
-      void operator()(const double_var_decl& x) const {
-        std::vector<expression> read_args;
-        generate_initialize_array_bounded(x,
-                                          is_var_context_ ? "T__" : "double",
-                                          "scalar", read_args);
-      }
-      void operator()(const vector_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.M_);
-        generate_initialize_array_bounded(x,
-                                          is_var_context_
-                                          ? "Eigen::Matrix"
-                                          "<T__,Eigen::Dynamic,1> "
-                                          : "vector_d",
-                                          "vector", read_args);
-      }
-      void operator()(const row_vector_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.N_);
-        generate_initialize_array_bounded(x,
-                                          is_var_context_
-                                          ? "Eigen::Matrix"
-                                          "<T__,1,Eigen::Dynamic> "
-                                          : "row_vector_d",
-                                          "row_vector", read_args);
-      }
-      void operator()(const matrix_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.M_);
-        read_args.push_back(x.N_);
-        generate_initialize_array_bounded(x,
-                                          is_var_context_
-                                          ? "Eigen::Matrix"
-                                          "<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                                          : "matrix_d",
-                                          "matrix", read_args);
-      }
-      void operator()(const unit_vector_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.K_);
-        generate_initialize_array(is_var_context_
-                                  ? "Eigen::Matrix"
-                                  "<T__,Eigen::Dynamic,1> "
-                                  : "vector_d",
-                                  "unit_vector", read_args, x.name_, x.dims_);
-      }
-      void operator()(const simplex_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.K_);
-        generate_initialize_array(is_var_context_
-                                  ? "Eigen::Matrix"
-                                  "<T__,Eigen::Dynamic,1> "
-                                  : "vector_d",
-                                  "simplex", read_args, x.name_, x.dims_);
-      }
-      void operator()(const ordered_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.K_);
-        generate_initialize_array(is_var_context_
-                                  ? "Eigen::Matrix"
-                                  "<T__,Eigen::Dynamic,1> "
-                                  : "vector_d",
-                                  "ordered", read_args, x.name_, x.dims_);
-      }
-      void operator()(const positive_ordered_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.K_);
-        generate_initialize_array(is_var_context_
-                                  ? "Eigen::Matrix"
-                                  "<T__,Eigen::Dynamic,1> "
-                                  : "vector_d",
-                                  "positive_ordered", read_args,
-                                  x.name_, x.dims_);
-      }
-      void operator()(const cholesky_factor_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.M_);
-        read_args.push_back(x.N_);
-        generate_initialize_array(is_var_context_
-                                  ? "Eigen::Matrix"
-                                  "<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                                  : "matrix_d",
-                                  "cholesky_factor", read_args,
-                                  x.name_, x.dims_);
-      }
-      void operator()(const cholesky_corr_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.K_);
-        generate_initialize_array(is_var_context_
-                                  ? "Eigen::Matrix"
-                                  "<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                                  : "matrix_d",
-                                  "cholesky_corr", read_args, x.name_, x.dims_);
-      }
-
-      void operator()(const cov_matrix_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.K_);
-        generate_initialize_array(is_var_context_
-                                  ? "Eigen::Matrix"
-                                  "<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                                  : "matrix_d",
-                                  "cov_matrix", read_args, x.name_, x.dims_);
-      }
-      void operator()(const corr_matrix_var_decl& x) const {
-        std::vector<expression> read_args;
-        read_args.push_back(x.K_);
-        generate_initialize_array(is_var_context_
-                                  ? "Eigen::Matrix"
-                                  "<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                                  : "matrix_d",
-                                  "corr_matrix", read_args, x.name_, x.dims_);
-      }
-
-
-      void generate_initialize_array(const std::string& var_type,
-                                     const std::string& read_type,
-                                     const std::vector<expression>& read_args,
-                                     const std::string& name,
-                                     const std::vector<expression>& dims)
-      const {
-        if (declare_vars_) {
-          o_ << INDENT2;
-          for (size_t i = 0; i < dims.size(); ++i) o_ << "vector<";
-          o_ << var_type;
-          for (size_t i = 0; i < dims.size(); ++i) o_ << "> ";
-          if (dims.size() == 0) o_ << " ";
-          o_ << name << ";" << EOL;
-        }
-
-        if (dims.size() == 0) {
-          generate_void_statement(name, 2, o_);
-          o_ << INDENT2 << "if (jacobian__)" << EOL;
-
-          // w Jacobian
-          generate_indent(3, o_);
-          o_ << name << " = in__." << read_type  << "_constrain(";
-          for (size_t j = 0; j < read_args.size(); ++j) {
-            if (j > 0) o_ << ",";
-            generate_expression(read_args[j], o_);
-          }
-          if (read_args.size() > 0)
-            o_ << ",";
-          o_ << "lp__";
-          o_ << ");" << EOL;
-
-          o_ << INDENT2 << "else" << EOL;
-
-          // w/o Jacobian
-          generate_indent(3, o_);
-          o_ << name << " = in__." << read_type  << "_constrain(";
-          for (size_t j = 0; j < read_args.size(); ++j) {
-            if (j > 0) o_ << ",";
-            generate_expression(read_args[j], o_);
-          }
-          o_ << ");" << EOL;
-
-        } else {
-          // dims > 0
-          std::string name_dims(name);
-          for (size_t i = 0; i < dims.size(); ++i) {
-            generate_indent(i + 2, o_);
-            o_ << "size_t dim_"  << name << "_" << i << "__ = ";
-            generate_expression(dims[i], o_);
-            o_ << ";" << EOL;
-
-            if (i < dims.size() - 1) {
-              generate_indent(i + 2, o_);
-              o_ << name_dims << ".resize(dim" << "_"
-                 << name << "_" << i << "__);"
-                 << EOL;
-              name_dims.append("[k_").append(to_string(i)).append("__]");
-            }
-
-            generate_indent(i + 2, o_);
-            if (i == dims.size() - 1) {
-              o_ << name_dims << ".reserve(dim_" << name
-                 << "_" << i << "__);" << EOL;
-              generate_indent(i + 2, o_);
-            }
-
-            o_ << "for (size_t k_" << i << "__ = 0;"
-               << " k_" << i << "__ < dim_" << name << "_" << i << "__;"
-               << " ++k_" << i << "__) {" << EOL;
-
-            // if on the last loop, push read element into array
-            if (i == dims.size() - 1) {
-              generate_indent(i + 3, o_);
-              o_ << "if (jacobian__)" << EOL;
-
-              // w Jacobian
-              generate_indent(i + 4, o_);
-              o_ << name_dims << ".push_back(in__."
-                 << read_type << "_constrain(";
-              for (size_t j = 0; j < read_args.size(); ++j) {
-                if (j > 0) o_ << ",";
-                generate_expression(read_args[j], o_);
-              }
-              if (read_args.size() > 0)
-                o_ << ",";
-              o_ << "lp__";
-              o_ << "));" << EOL;
-
-              generate_indent(i + 3, o_);
-              o_ << "else" << EOL;
-
-              // w/o Jacobian
-              generate_indent(i + 4, o_);
-              o_ << name_dims << ".push_back(in__."
-                 << read_type << "_constrain(";
-              for (size_t j = 0; j < read_args.size(); ++j) {
-                if (j > 0) o_ << ",";
-                generate_expression(read_args[j], o_);
-              }
-              o_ << "));" << EOL;
-            }
-          }
-
-          for (size_t i = dims.size(); i > 0; --i) {
-            generate_indent(i + 1, o_);
-            o_ << "}" << EOL;
-          }
-        }
-        o_ << EOL;
-      }
-    };
-
-
-    void generate_local_var_inits(std::vector<var_decl> vs,
-                                  bool is_var_context,
-                                  bool declare_vars,
-                                  std::ostream& o) {
-      o << INDENT2
-        << "stan::io::reader<"
-        << (is_var_context ? "T__" : "double")
-        << "> in__(params_r__,params_i__);" << EOL2;
-      init_local_var_visgen vis_init(declare_vars, is_var_context, o);
-      for (size_t i = 0; i < vs.size(); ++i) {
-          boost::apply_visitor(vis_init, vs[i].decl_);
-      }
-    }
-
-
-    void generate_public_decl(std::ostream& o) {
-      o << "public:" << EOL;
-    }
-
-    void generate_private_decl(std::ostream& o) {
-      o << "private:" << EOL;
-    }
-
-    struct validate_var_decl_visgen : public visgen {
-      int indents_;
-      validate_var_decl_visgen(int indents,
-                               std::ostream& o)
-        : visgen(o),
-          indents_(indents) {
-      }
-      void generate_begin_for_dims(const std::vector<expression>& dims)
-        const {
-        for (size_t i = 0; i < dims.size(); ++i) {
-          generate_indent(indents_+i, o_);
-          o_ << "for (int k" << i << "__ = 0;"
-             << " k" << i << "__ < ";
-          generate_expression(dims[i].expr_, o_);
-          o_ << ";";
-          o_ << " ++k" << i << "__) {" << EOL;
-        }
-      }
-      void generate_end_for_dims(size_t dims_size) const {
-        for (size_t i = 0; i < dims_size; ++i) {
-          generate_indent(indents_ + dims_size - i - 1, o_);
-          o_ << "}" << EOL;
-        }
-      }
-      void generate_loop_var(const std::string& name,
-                             size_t dims_size) const {
-        o_ << name;
-        for (size_t i = 0; i < dims_size; ++i)
-          o_ << "[k" << i << "__]";
-      }
-      void operator()(const nil& /*x*/) const { }
-      template <typename T>
-      void basic_validate(const T& x) const {
-        if (!(x.range_.has_low() || x.range_.has_high())) {
-          return;  // unconstrained
-        }
-        generate_begin_for_dims(x.dims_);
-        if (x.range_.has_low()) {
-          generate_indent(indents_ + x.dims_.size(), o_);
-          o_ << "check_greater_or_equal(function__,";
-          o_ << "\"";
-          generate_loop_var(x.name_, x.dims_.size());
-          o_ << "\",";
-          generate_loop_var(x.name_, x.dims_.size());
-          o_ << ",";
-          generate_expression(x.range_.low_.expr_, o_);
-          o_ << ");" << EOL;
-        }
-        if (x.range_.has_high()) {
-          generate_indent(indents_ + x.dims_.size(), o_);
-          o_ << "check_less_or_equal(function__,";
-          o_ << "\"";
-          generate_loop_var(x.name_, x.dims_.size());
-          o_ << "\",";
-          generate_loop_var(x.name_, x.dims_.size());
-          o_ << ",";
-          generate_expression(x.range_.high_.expr_, o_);
-          o_ << ");" << EOL;
-        }
-        generate_end_for_dims(x.dims_.size());
-      }
-      void operator()(const int_var_decl& x) const {
-        basic_validate(x);
-      }
-      void operator()(const double_var_decl& x) const {
-        basic_validate(x);
-      }
-      void operator()(const vector_var_decl& x) const {
-        basic_validate(x);
-      }
-      void operator()(const row_vector_var_decl& x) const {
-        basic_validate(x);
-      }
-      void operator()(const matrix_var_decl& x) const {
-        basic_validate(x);
-      }
-      template <typename T>
-      void nonbasic_validate(const T& x,
-                             const std::string& type_name) const {
-        generate_begin_for_dims(x.dims_);
-        generate_indent(indents_ + x.dims_.size(), o_);
-        o_ << "stan::math::check_" << type_name << "(function__,";
-        o_ << "\"";
-        generate_loop_var(x.name_, x.dims_.size());
-        o_ << "\",";
-        generate_loop_var(x.name_, x.dims_.size());
-        o_ << ");"
-           << EOL;
-        generate_end_for_dims(x.dims_.size());
-      }
-      void operator()(const unit_vector_var_decl& x) const {
-        nonbasic_validate(x, "unit_vector");
-      }
-      void operator()(const simplex_var_decl& x) const {
-        nonbasic_validate(x, "simplex");
-      }
-      void operator()(const ordered_var_decl& x) const {
-        nonbasic_validate(x, "ordered");
-      }
-      void operator()(const positive_ordered_var_decl& x) const {
-        nonbasic_validate(x, "positive_ordered");
-      }
-      void operator()(const cholesky_factor_var_decl& x) const {
-        nonbasic_validate(x, "cholesky_factor");
-      }
-      void operator()(const cholesky_corr_var_decl& x) const {
-        nonbasic_validate(x, "cholesky_factor_corr");
-      }
-      void operator()(const cov_matrix_var_decl& x) const {
-        nonbasic_validate(x, "cov_matrix");
-      }
-      void operator()(const corr_matrix_var_decl& x) const {
-        nonbasic_validate(x, "corr_matrix");
-      }
-    };
-
-
-    void generate_validate_var_decl(const var_decl& decl,
-                                    int indent,
-                                    std::ostream& o) {
-      validate_var_decl_visgen vis(indent, o);
-      boost::apply_visitor(vis, decl.decl_);
-    }
-
-    void generate_validate_var_decls(const std::vector<var_decl> decls,
-                                     int indent,
-                                     std::ostream& o) {
-      for (size_t i = 0; i < decls.size(); ++i)
-        generate_validate_var_decl(decls[i], indent, o);
-    }
-
-    // see _var_decl_visgen cut & paste
-    struct member_var_decl_visgen : public visgen {
-      int indents_;
-      member_var_decl_visgen(int indents,
-                             std::ostream& o)
-        : visgen(o),
-          indents_(indents) {
-      }
-      void operator()(const nil& /*x*/) const { }
-      void operator()(const int_var_decl& x) const {
-        declare_array("int", x.name_, x.dims_.size());
-      }
-      void operator()(const double_var_decl& x) const {
-        declare_array("double", x.name_, x.dims_.size());
-      }
-      void operator()(const unit_vector_var_decl& x) const {
-        declare_array(("vector_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const simplex_var_decl& x) const {
-        declare_array(("vector_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const ordered_var_decl& x) const {
-        declare_array(("vector_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const positive_ordered_var_decl& x) const {
-        declare_array(("vector_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const cholesky_factor_var_decl& x) const {
-        declare_array(("matrix_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const cholesky_corr_var_decl& x) const {
-        declare_array(("matrix_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const cov_matrix_var_decl& x) const {
-        declare_array(("matrix_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const corr_matrix_var_decl& x) const {
-        declare_array(("matrix_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const vector_var_decl& x) const {
-        declare_array(("vector_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const row_vector_var_decl& x) const {
-        declare_array(("row_vector_d"), x.name_, x.dims_.size());
-      }
-      void operator()(const matrix_var_decl& x) const {
-        declare_array(("matrix_d"), x.name_, x.dims_.size());
-      }
-      void declare_array(const std::string& type,
-                         const std::string& name,
-                         size_t size) const {
-        for (int i = 0; i < indents_; ++i)
-          o_ << INDENT;
-        for (size_t i = 0; i < size; ++i) {
-          o_ << "vector<";
-        }
-        o_ << type;
-        if (size > 0) {
-          o_ << ">";
-        }
-        for (size_t i = 1; i < size; ++i) {
-          o_ << " >";
-        }
-        o_ << " " << name << ";" << EOL;
-      }
-    };
-
-    void generate_member_var_decls(const std::vector<var_decl>& vs,
-                                   int indent,
-                                   std::ostream& o) {
-      member_var_decl_visgen vis(indent, o);
-      for (size_t i = 0; i < vs.size(); ++i)
-        boost::apply_visitor(vis, vs[i].decl_);
-    }
-
-    /*
-     * generates local data variable declaration
-     */
-    struct local_var_decl_visgen : public visgen {
-      int indents_;
-      bool is_var_context_;
-      bool is_fun_return_;
-      explicit local_var_decl_visgen(int indents,
-                            bool is_var_context,
-                            bool is_fun_return,
-                            std::ostream& o)
-        : visgen(o),
-          indents_(indents),
-          is_var_context_(is_var_context),
-          is_fun_return_(is_fun_return) {
-      }
-      void operator()(const nil& /*x*/) const { }
-      void operator()(const int_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        declare_array("int", ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const double_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        declare_array(is_fun_return_
-                      ? "fun_scalar_t__"
-                      : (is_var_context_ ? "T__" : "double"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const vector_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.M_, indents_, o_);
-        ctor_args.push_back(x.M_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,Eigen::Dynamic,1> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,1> " : "vector_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const row_vector_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.N_, indents_, o_);
-        ctor_args.push_back(x.N_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,1,Eigen::Dynamic> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,1,Eigen::Dynamic> "
-                         : "row_vector_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const matrix_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.M_, indents_, o_);
-        generate_validate_positive(x.name_, x.N_, indents_, o_);
-        ctor_args.push_back(x.M_);
-        ctor_args.push_back(x.N_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,"
-                      "Eigen::Dynamic,Eigen::Dynamic> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                         : "matrix_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const unit_vector_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.K_, indents_, o_);
-        ctor_args.push_back(x.K_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,Eigen::Dynamic,1> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,1> " : "vector_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const simplex_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.K_, indents_, o_);
-        ctor_args.push_back(x.K_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,Eigen::Dynamic,1> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,1> " : "vector_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const ordered_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.K_, indents_, o_);
-        ctor_args.push_back(x.K_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,Eigen::Dynamic,1> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,1> " : "vector_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const positive_ordered_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.K_, indents_, o_);
-        ctor_args.push_back(x.K_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,Eigen::Dynamic,1> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,1> " : "vector_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const cholesky_factor_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.M_, indents_, o_);
-        generate_validate_positive(x.name_, x.N_, indents_, o_);
-        ctor_args.push_back(x.M_);
-        ctor_args.push_back(x.N_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,"
-                      "Eigen::Dynamic,Eigen::Dynamic> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                         : "matrix_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const cholesky_corr_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.K_, indents_, o_);
-        ctor_args.push_back(x.K_);
-        ctor_args.push_back(x.K_);
-        declare_array(is_var_context_
-                      ? "Eigen::Matrix<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                      : "matrix_d",
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const cov_matrix_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.K_, indents_, o_);
-        ctor_args.push_back(x.K_);
-        ctor_args.push_back(x.K_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,"
-                      "Eigen::Dynamic,Eigen::Dynamic> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                         : "matrix_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void operator()(const corr_matrix_var_decl& x) const {
-        std::vector<expression> ctor_args;
-        generate_validate_positive(x.name_, x.K_, indents_, o_);
-        ctor_args.push_back(x.K_);
-        ctor_args.push_back(x.K_);
-        declare_array(is_fun_return_
-                      ? "Eigen::Matrix<fun_scalar_t__,"
-                      "Eigen::Dynamic,Eigen::Dynamic> "
-                      : (is_var_context_
-                         ? "Eigen::Matrix<T__,Eigen::Dynamic,Eigen::Dynamic> "
-                         : "matrix_d"),
-                      ctor_args, x.name_, x.dims_);
-      }
-      void generate_type(const std::string& type,
-                         size_t num_dims) const {
-        for (size_t i = 0; i < num_dims; ++i)
-          o_ << "vector<";
-        o_ << type;
-        for (size_t i = 0; i < num_dims; ++i) {
-          if (i > 0) o_ << " ";
-          o_ << ">";
-        }
-      }
-      void generate_void_statement(const std::string& name) const {
-        o_ << "(void) " << name << ";  // dummy to suppress unused var warning";
-      }
-      void generate_init_args(const std::string& type,
-                              const std::vector<expression>& ctor_args,
-                              const std::vector<expression>& dims,
-                              size_t dim) const {
-        if (dim < dims.size()) {  // more dims left
-          o_ << '(';  // open(1)
-          generate_expression(dims[dim], o_);
-          if ((dim + 1 < dims.size()) ||  ctor_args.size() > 0) {
-            o_ << ", (";  // open(2)
-            generate_type(type, dims.size() - dim - 1);
-            generate_init_args(type, ctor_args, dims, dim + 1);
-            o_ << ')';  // close(2)
-          } else if (type == "var") {
-            o_ << ", DUMMY_VAR__";
-          } else if (type == "int") {
-            o_ << ", 0";
-          } else if (type == "double") {
-            o_ << ", 0.0";
-          } else {
-            // shouldn't hit this
-          }
-          o_ << ')';  // close(1)
-        } else {
-          if (ctor_args.size() == 0) {  // scalar int or real
-            if (type == "int") {
-              o_ << "(0)";
-            } else if (type == "double") {
-              o_ << "(0.0)";
-            } else if (type == "var") {
-              o_ << "(DUMMY_VAR__)";
-            } else {
-              // shouldn't hit this, either
-            }
-          } else if (ctor_args.size() == 1) {  // vector
-            o_ << '(';
-            generate_eigen_index_expression(ctor_args[0], o_);
-            o_ << ')';
-          } else if (ctor_args.size() > 1) {  // matrix
-            o_ << '(';
-            generate_eigen_index_expression(ctor_args[0], o_);
-            o_ << ',';
-            generate_eigen_index_expression(ctor_args[1], o_);
-            o_ << ')';
-          }
-        }
-      }
-      void declare_array(const std::string& type,
-                         const std::vector<expression>& ctor_args,
-                         const std::string& name,
-                         const std::vector<expression>& dims,
-                         const expression& definition = expression()) const {
-        // check array dimensions
-        for (size_t i = 0; i < dims.size(); ++i)
-          generate_validate_positive(name, dims[i], indents_, o_);
-        // require double parens to counter "most vexing parse" problem
-        generate_indent(indents_, o_);
-        generate_type(type, dims.size());
-        o_ << ' '  << name;
-        generate_init_args(type, ctor_args, dims, 0);
-        o_ << ";" << EOL;
-        if (dims.size() == 0) {
-          generate_indent(indents_, o_);
-          generate_void_statement(name);
-          o_ << EOL;
-        }
-        if (type == "Eigen::Matrix<T__, Eigen::Dynamic, Eigen::Dynamic> "
-            || type == "Eigen::Matrix<T__, 1, Eigen::Dynamic> "
-            || type == "Eigen::Matrix<T__, Eigen::Dynamic, 1> ") {
-          generate_indent(indents_, o_);
-          o_ << "stan::math::fill(" << name << ", DUMMY_VAR__);" << EOL;
-        }
-      }
-    };
-
-    void generate_local_var_decls(const std::vector<var_decl>& vs,
-                                  int indent,
-                                  std::ostream& o,
-                                  bool is_var_context,
-                                  bool is_fun_return) {
-      local_var_decl_visgen vis_decl(indent, is_var_context, is_fun_return, o);
-      local_var_init_nan_visgen vis_init(is_var_context, indent, o);
-      init_vars_visgen vis_filler(indent, o);
-      for (size_t i = 0; i < vs.size(); ++i) {
-        boost::apply_visitor(vis_decl, vs[i].decl_);
-        boost::apply_visitor(vis_init, vs[i].decl_);
-        boost::apply_visitor(vis_filler, vs[i].decl_);
-        if (vs[i].has_def()) {
-          generate_indent(indent, o);
-          o << "stan::math::assign("
-            << vs[i].name()
-            << ",";
-          generate_expression(vs[i].def(), false, is_var_context, o);
-          o << ");" << EOL;
-        }
-      }
-      o << EOL;
-    }
-
-
-    void generate_local_var_init_nan(const std::vector<var_decl>& vs,
-                                     int indent,
-                                     std::ostream& o,
-                                     bool is_var_context,
-                                     bool is_fun_return) {
-      local_var_init_nan_visgen vis(is_var_context, indent, o);
-      for (size_t i = 0; i < vs.size(); ++i) {
-        boost::apply_visitor(vis, vs[i].decl_);
-      }
-    }
-
-
-    void generate_define_vars(const std::vector<var_decl>& vs,
-                              int indent,
-                              bool is_var_context,
-                              std::ostream& o) {
-      generate_comment("assign variable definitions",
-                       indent, o);
-      for (size_t i = 0; i < vs.size(); ++i) {
-        if (vs[i].has_def()) {
-          generate_indent(indent, o);
-          o << "stan::math::assign("
-            << vs[i].name()
-            << ",";
-          generate_expression(vs[i].def(), false, is_var_context, o);
-          o << ");" << EOL;
-        }
-      }
-    }
-
-    struct validate_transformed_params_visgen : public visgen {
-      int indents_;
-      validate_transformed_params_visgen(int indents,
-                                         std::ostream& o)
-        : visgen(o),
-          indents_(indents)
-      { }
-      void operator()(const nil& /*x*/) const { }
-      void operator()(const int_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        validate_array(x.name_, dims, 0);
-      }
-      void operator()(const double_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        validate_array(x.name_, dims, 0);
-      }
-      void operator()(const vector_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.M_);
-        validate_array(x.name_, dims, 1);
-      }
-      void operator()(const unit_vector_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.K_);
-        validate_array(x.name_, dims, 1);
-      }
-      void operator()(const simplex_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.K_);
-        validate_array(x.name_, dims, 1);
-      }
-      void operator()(const ordered_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.K_);
-        validate_array(x.name_, dims, 1);
-      }
-      void operator()(const positive_ordered_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.K_);
-        validate_array(x.name_, dims, 1);
-      }
-      void operator()(const row_vector_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.N_);
-        validate_array(x.name_, dims, 1);
-      }
-      void operator()(const matrix_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.M_);
-        dims.push_back(x.N_);
-        validate_array(x.name_, dims, 2);
-      }
-      void operator()(const cholesky_factor_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.M_);
-        dims.push_back(x.N_);
-        validate_array(x.name_, dims, 2);
-      }
-      void operator()(const cholesky_corr_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.K_);
-        dims.push_back(x.K_);
-        validate_array(x.name_, dims, 2);
-      }
-      void operator()(const cov_matrix_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.K_);
-        dims.push_back(x.K_);
-        validate_array(x.name_, dims, 2);
-      }
-      void operator()(const corr_matrix_var_decl& x) const {
-        std::vector<expression> dims(x.dims_);
-        dims.push_back(x.K_);
-        dims.push_back(x.K_);
-        validate_array(x.name_, dims, 2);
-      }
-      void validate_array(const std::string& name,
-                          const std::vector<expression>& dims,
-                          size_t matrix_dims) const {
-        size_t non_matrix_dims = dims.size() - matrix_dims;
-
-        for (size_t k = 0; k < dims.size(); ++k) {
-          generate_indent(indents_ + k, o_);
-          o_ << "for (int i" << k << "__ = 0; i" << k << "__ < ";
-          generate_expression(dims[k], o_);
-          o_ << "; ++i" << k << "__) {" << EOL;
-        }
-
-        generate_indent(indents_ + dims.size(), o_);
-        o_ << "if (stan::math::is_uninitialized(" << name;
-        for (size_t k = 0; k < non_matrix_dims; ++k)
-          o_ << "[i" << k << "__]";
-        if (matrix_dims > 0) {
-          o_ << "(i" << non_matrix_dims << "__";
-          if (matrix_dims > 1)
-            o_ << ",i" << (non_matrix_dims + 1) << "__";
-          o_ << ')';
-        }
-        o_ << ")) {" << EOL;
-        generate_indent(indents_ + dims.size() + 1, o_);
-        o_ << "std::stringstream msg__;" << EOL;
-        generate_indent(indents_ + dims.size() + 1, o_);
-        o_ << "msg__ << \"Undefined transformed parameter: "
-           << name << "\"";
-        for (size_t k = 0; k < dims.size(); ++k) {
-          o_ << " << '['";
-          o_ << " << i" << k << "__";
-          o_ << " << ']'";
-        }
-        o_ << ';' << EOL;
-        generate_indent(indents_ + dims.size() + 1, o_);
-        o_ << "throw std::runtime_error(msg__.str());" << EOL;
-
-        generate_indent(indents_ + dims.size(), o_);
-        o_ << "}" << EOL;
-        for (size_t k = 0; k < dims.size(); ++k) {
-          generate_indent(indents_ + dims.size() - k - 1, o_);
-          o_ << "}" << EOL;
-        }
-      }
-    };
-
-    void generate_validate_transformed_params(const std::vector<var_decl>& vs,
-                                              int indent,
-                                              std::ostream& o) {
-      generate_comment("validate transformed parameters", indent, o);
-      validate_transformed_params_visgen vis(indent, o);
-      for (size_t i = 0; i < vs.size(); ++i)
-        boost::apply_visitor(vis, vs[i].decl_);
-      o << EOL;
-    }
-
-    struct idx_visgen : public visgen {
-      explicit idx_visgen(std::ostream& o): visgen(o) { }
-      void operator()(const uni_idx& i) const {
-        o_ << "stan::model::index_uni(";
-        generate_expression(i.idx_, o_);
-        o_ << ")";
-      }
-      void operator()(const multi_idx& i) const {
-        o_ << "stan::model::index_multi(";
-        generate_expression(i.idxs_, o_);
-        o_ << ")";
-      }
-      void operator()(const omni_idx& i) const {
-        o_ << "stan::model::index_omni()";
-      }
-      void operator()(const lb_idx& i) const {
-        o_ << "stan::model::index_min(";
-        generate_expression(i.lb_, o_);
-        o_ << ")";
-      }
-      void operator()(const ub_idx& i) const {
-        o_ << "stan::model::index_max(";
-        generate_expression(i.ub_, o_);
-        o_ << ")";
-      }
-      void operator()(const lub_idx& i) const {
-        o_ << "stan::model::index_min_max(";
-        generate_expression(i.lb_, o_);
-        o_ << ", ";
-        generate_expression(i.ub_, o_);
-        o_ << ")";
-      }
-    };
-
-    void generate_idx(const idx& i, std::ostream& o) {
-      idx_visgen vis(o);
-      boost::apply_visitor(vis, i.idx_);
-    }
-
-    void generate_idxs(size_t pos, const std::vector<idx>& idxs,
-                       std::ostream& o) {
-      if (pos == idxs.size()) {
-        o << "stan::model::nil_index_list()";
-      } else {
-        o << "stan::model::cons_list(";
-        generate_idx(idxs[pos], o);
-        o << ", ";
-        generate_idxs(pos + 1, idxs, o);
-        o << ")";
-      }
-    }
-
-    void generate_idxs(const std::vector<idx>& idxs, std::ostream& o) {
-      generate_idxs(0, idxs, o);
-    }
-
-
-    struct idx_user_visgen : public visgen {
-      explicit idx_user_visgen(std::ostream& o): visgen(o) { }
-      void operator()(const uni_idx& i) const {
-        generate_expression(i.idx_, true, o_);
-      }
-      void operator()(const multi_idx& i) const {
-        generate_expression(i.idxs_, true, o_);
-      }
-      void operator()(const omni_idx& i) const {
-        o_ << " ";
-      }
-      void operator()(const lb_idx& i) const {
-        generate_expression(i.lb_, true, o_);
-        o_ << ": ";
-      }
-      void operator()(const ub_idx& i) const {
-        o_ << " :";
-        generate_expression(i.ub_, true, o_);
-      }
-      void operator()(const lub_idx& i) const {
-        generate_expression(i.lb_, true, o_);
-        o_ << ":";
-        generate_expression(i.ub_, true, o_);
-      }
-    };
-
-    void generate_idx_user(const idx& i, std::ostream& o) {
-      idx_user_visgen vis(o);
-      boost::apply_visitor(vis, i.idx_);
-    }
-
-    void generate_idxs_user(const std::vector<idx>& idxs, std::ostream& o) {
-      if (idxs.size() == 0)
-        return;
-      o << "[";
-      for (size_t i = 0; i < idxs.size(); ++i) {
-        if (i > 0)
-          o << ", ";
-        generate_idx_user(idxs[i], o);
-      }
-      o << "]";
-    }
-
-    struct statement_visgen : public visgen {
-      size_t indent_;
-      bool include_sampling_;
-      bool is_var_context_;
-      bool is_fun_return_;
-      statement_visgen(size_t indent,
-                       bool include_sampling,
-                       bool is_var_context,
-                       bool is_fun_return,
-                       std::ostream& o)
-        : visgen(o),
-          indent_(indent),
-          include_sampling_(include_sampling),
-          is_var_context_(is_var_context),
-          is_fun_return_(is_fun_return) {
-      }
-      void operator()(const nil& /*x*/) const {
-      }
-      void operator()(const assignment& x) const {
-        generate_indent(indent_, o_);
-        o_ << "stan::math::assign(";
-        generate_indexed_expr<true>(x.var_dims_.name_,
-                                    x.var_dims_.dims_,
-                                    x.var_type_.base_type_,
-                                    x.var_type_.dims_.size(),
-                                    false,
-                                    o_);
-        o_ << ", ";
-        generate_expression(x.expr_, false, is_var_context_, o_);
-        o_ << ");" << EOL;
-      }
-      void operator()(const assgn& y) const {
-        generate_indent(indent_, o_);
-        o_ << "stan::model::assign(";
-
-        expression var_expr(y.lhs_var_);
-        generate_expression(var_expr, false, is_var_context_, o_);
-        o_ << ", "
-           << EOL;
-
-        generate_indent(indent_ + 3, o_);
-        generate_idxs(y.idxs_, o_);
-        o_ << ", "
-           << EOL;
-
-        generate_indent(indent_ + 3, o_);
-        if (y.lhs_var_occurs_on_rhs()) {
-          o_ << "stan::model::deep_copy(";
-          generate_expression(y.rhs_, false, is_var_context_, o_);
-          o_ << ")";
-        } else {
-          generate_expression(y.rhs_, false, is_var_context_, o_);
-        }
-
-        o_ << ", "
-           << EOL;
-        generate_indent(indent_ + 3, o_);
-        o_ << '"'
-           << "assigning variable "
-           << y.lhs_var_.name_
-           << '"';
-        o_ << ");"
-           << EOL;
-      }
-      void operator()(const expression& x) const {
-        generate_indent(indent_, o_);
-        generate_expression(x, false, is_var_context_, o_);
-        o_ << ";" << EOL;
-      }
-      void operator()(const sample& x) const {
-        if (!include_sampling_) return;
-        std::string prob_fun = get_prob_fun(x.dist_.family_);
-        generate_indent(indent_, o_);
-        o_ << "lp_accum__.add(" << prob_fun << "<propto__>(";
-        generate_expression(x.expr_, o_);
-        for (size_t i = 0; i < x.dist_.args_.size(); ++i) {
-          o_ << ", ";
-          generate_expression(x.dist_.args_[i], o_);
-        }
-        bool is_user_defined
-          = is_user_defined_prob_function(prob_fun, x.expr_, x.dist_.args_);
-        if (is_user_defined)
-          o_ << ", pstream__";
-        o_ << "));" << EOL;
-        // rest of impl is for truncation
-        // test variable is within truncation interval
-        if (x.truncation_.has_low()) {
-          generate_indent(indent_, o_);
-          o_ << "if (";
-          generate_expression(x.expr_,  o_);
-          o_ << " < ";
-          generate_expression(x.truncation_.low_.expr_, o_);
-          o_ << ") lp_accum__.add(-std::numeric_limits<double>::infinity());"
-             << EOL;
-        }
-        if (x.truncation_.has_high()) {
-          generate_indent(indent_, o_);
-          if (x.truncation_.has_low()) o_ << "else ";
-          o_ << "if (";
-          generate_expression(x.expr_, o_);
-          o_ << " > ";
-          generate_expression(x.truncation_.high_.expr_, o_);
-          o_ << ") lp_accum__.add(-std::numeric_limits<double>::infinity());"
-             << EOL;
-        }
-        // generate log denominator for case where bounds test pass
-        if (x.truncation_.has_low() || x.truncation_.has_high())
-          generate_truncation(x, is_user_defined, prob_fun);
-      }
-
-      void generate_truncation(const sample& x, bool is_user_defined,
-                               const std::string& prob_fun) const {
-        std::stringstream sso_lp;
-        generate_indent(indent_, o_);
-        if (x.truncation_.has_low() && x.truncation_.has_high()) {
-          // T[L,U]: -log_diff_exp(Dist_cdf_log(U|params),
-          //                       Dist_cdf_log(L|Params))
-          sso_lp << "log_diff_exp(";
-          sso_lp << get_cdf(x.dist_.family_) << "(";
-          generate_expression(x.truncation_.high_.expr_, sso_lp);
-          for (size_t i = 0; i < x.dist_.args_.size(); ++i) {
-            sso_lp << ", ";
-            generate_expression(x.dist_.args_[i], sso_lp);
-          }
-          if (is_user_defined)
-            sso_lp << ", pstream__";
-          sso_lp << "), " << get_cdf(x.dist_.family_) << "(";
-          generate_expression(x.truncation_.low_.expr_, sso_lp);
-          for (size_t i = 0; i < x.dist_.args_.size(); ++i) {
-            sso_lp << ", ";
-            generate_expression(x.dist_.args_[i], sso_lp);
-          }
-          if (is_user_defined)
-            sso_lp << ", pstream__";
-          sso_lp << "))";
-
-        } else if (!x.truncation_.has_low() && x.truncation_.has_high()) {
-          // T[,U];  -Dist_cdf_log(U)
-          sso_lp << get_cdf(x.dist_.family_) << "(";
-          generate_expression(x.truncation_.high_.expr_, sso_lp);
-          for (size_t i = 0; i < x.dist_.args_.size(); ++i) {
-            sso_lp << ", ";
-            generate_expression(x.dist_.args_[i], sso_lp);
-          }
-          if (is_user_defined)
-            sso_lp << ", pstream__";
-          sso_lp << ")";
-
-        } else if (x.truncation_.has_low() && !x.truncation_.has_high()) {
-          // T[L,]: -Dist_ccdf_log(L)
-          sso_lp << get_ccdf(x.dist_.family_) << "(";
-          generate_expression(x.truncation_.low_.expr_, sso_lp);
-          for (size_t i = 0; i < x.dist_.args_.size(); ++i) {
-            sso_lp << ", ";
-            generate_expression(x.dist_.args_[i], sso_lp);
-          }
-          if (is_user_defined)
-            sso_lp << ", pstream__";
-          sso_lp << ")";
-        }
-
-        o_ << "else lp_accum__.add(-";
-
-        if (x.is_discrete() && x.truncation_.has_low()) {
-          o_ << "log_sum_exp(" << sso_lp.str() << ", ";
-          // generate adjustment for lower-bound off by 1 due to log CCDF
-          o_ << prob_fun << "(";
-          generate_expression(x.truncation_.low_.expr_, o_);
-          for (size_t i = 0; i < x.dist_.args_.size(); ++i) {
-            o_ << ", ";
-            generate_expression(x.dist_.args_[i], o_);
-          }
-          if (is_user_defined) o_ << ", pstream__";
-          o_ << "))";
-        } else {
-          o_ << sso_lp.str();
-        }
-
-        o_ << ");" << std::endl;
-      }
-
-      void operator()(const increment_log_prob_statement& x) const {
-        generate_indent(indent_, o_);
-        o_ << "lp_accum__.add(";
-        generate_expression(x.log_prob_, o_);
-        o_ << ");" << EOL;
-      }
-
-      void operator()(const statements& x) const {
-        bool has_local_vars = x.local_decl_.size() > 0;
-        size_t indent = has_local_vars ? (indent_ + 1) : indent_;
-        if (has_local_vars) {
-          generate_indent(indent_, o_);
-          o_ << "{" << EOL;
-          generate_local_var_decls(x.local_decl_, indent, o_,
-                                   is_var_context_, is_fun_return_);
-        }
-        o_ << EOL;
-        for (size_t i = 0; i < x.statements_.size(); ++i)
-          generate_statement(x.statements_[i], indent, o_, include_sampling_,
-                             is_var_context_, is_fun_return_);
-        if (has_local_vars) {
-          generate_indent(indent_, o_);
-          o_ << "}" << EOL;
-        }
-      }
-
-      void operator()(const print_statement& ps) const {
-        generate_indent(indent_, o_);
-        o_ << "if (pstream__) {" << EOL;
-        for (size_t i = 0; i < ps.printables_.size(); ++i) {
-          generate_indent(indent_ + 1, o_);
-          o_ << "stan_print(pstream__,";
-          generate_printable(ps.printables_[i], o_);
-          o_ << ");" << EOL;
-        }
-        generate_indent(indent_ + 1, o_);
-        o_ << "*pstream__ << std::endl;" << EOL;
-        generate_indent(indent_, o_);
-        o_ << '}' << EOL;
-      }
-      void operator()(const reject_statement& ps) const {
-        generate_indent(indent_, o_);
-        o_ << "std::stringstream errmsg_stream__;" << EOL;
-        for (size_t i = 0; i < ps.printables_.size(); ++i) {
-          generate_indent(indent_, o_);
-          o_ << "errmsg_stream__ << ";
-          generate_printable(ps.printables_[i], o_);
-          o_ << ";" << EOL;
-        }
-        generate_indent(indent_, o_);
-        o_ << "throw std::domain_error(errmsg_stream__.str());" << EOL;
-      }
-      void operator()(const return_statement& rs) const {
-        generate_indent(indent_, o_);
-        o_ << "return ";
-        if (!rs.return_value_.expression_type().is_ill_formed()
-            && !rs.return_value_.expression_type().is_void()) {
-          o_ << "stan::math::promote_scalar<fun_return_scalar_t__>(";
-          generate_expression(rs.return_value_, o_);
-          o_ << ")";
-        }
-        o_ << ";" << EOL;
-      }
-      void operator()(const for_statement& x) const {
-        generate_indent(indent_, o_);
-        o_ << "for (int " << x.variable_ << " = ";
-        generate_expression(x.range_.low_, o_);
-        o_ << "; " << x.variable_ << " <= ";
-        generate_expression(x.range_.high_, o_);
-        o_ << "; ++" << x.variable_ << ") {" << EOL;
-        generate_statement(x.statement_, indent_ + 1, o_, include_sampling_,
-                           is_var_context_, is_fun_return_);
-        generate_indent(indent_, o_);
-        o_ << "}" << EOL;
-      }
-      void operator()(const while_statement& x) const {
-        generate_indent(indent_, o_);
-        o_ << "while (as_bool(";
-        generate_expression(x.condition_, o_);
-        o_ << ")) {" << EOL;
-        generate_statement(x.body_, indent_+1, o_, include_sampling_,
-                           is_var_context_, is_fun_return_);
-        generate_indent(indent_, o_);
-        o_ << "}" << EOL;
-      }
-      void operator()(const break_continue_statement& st) const {
-        generate_indent(indent_, o_);
-        o_ << st.generate_ << ";" << EOL;
-      }
-      void operator()(const conditional_statement& x) const {
-        for (size_t i = 0; i < x.conditions_.size(); ++i) {
-          if (i == 0)
-            generate_indent(indent_, o_);
-          else
-            o_ << " else ";
-          o_ << "if (as_bool(";
-          generate_expression(x.conditions_[i], o_);
-          o_ << ")) {" << EOL;
-          generate_statement(x.bodies_[i], indent_ + 1, o_, include_sampling_,
-                             is_var_context_, is_fun_return_);
-          generate_indent(indent_, o_);
-          o_ << '}';
-        }
-        if (x.bodies_.size() > x.conditions_.size()) {
-          o_ << " else {" << EOL;
-          generate_statement(x.bodies_[x.bodies_.size()-1], indent_ + 1,
-                             o_, include_sampling_,
-                             is_var_context_, is_fun_return_);
-          generate_indent(indent_, o_);
-          o_ << '}';
-        }
-        o_ << EOL;
-      }
-      void operator()(const no_op_statement& /*x*/) const {
-      }
-    };
 
 
     struct is_numbered_statement_vis : public boost::static_visitor<bool> {
