+++ src/stan/lang/generator/constants.hpp
@@ -1,7 +1,9 @@
 #ifndef STAN_LANG_GENERATOR_CONSTANTS_HPP
 #define STAN_LANG_GENERATOR_CONSTANTS_HPP
 
+#include <stan/lang/ast.hpp>
 #include <string>
+#include <vector>
 
 namespace stan {
   namespace lang {
@@ -30,6 +32,11 @@ namespace stan {
      * Triple indentation.
      */
     const std::string INDENT3("            ");
+
+    /**
+     * Size zero vector of expressions.
+     */
+    const std::vector<expression> EMPTY_EXP_VECTOR(0);
 
   }
 }
